% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove_segdups.R
\name{check_in_segdup}
\alias{check_in_segdup}
\title{check if a de novo call is within the segdup regions}
\usage{
check_in_segdup(variant, segdups)
}
\arguments{
\item{variant}{dataframe/list for single variant}

\item{segdups}{dataframe of segmental duplication regions in the genome}
}
\value{
true/false for whwther the site is in  a segmental duplication.
}
\description{
check if a de novo call is within the segdup regions
}

