% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validations.R
\name{check_validations}
\alias{check_validations}
\title{check filtering performance against validated de novos}
\usage{
check_validations(de_novos, validations, plot_pp_dnm = FALSE)
}
\arguments{
\item{de_novos}{dataframe of candidate de novo sites}

\item{validations}{dataframe of de novo sites with validation status}
}
\value{
something
}
\description{
check filtering performance against validated de novos
}

