% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allele_counts.R
\name{count_gene_recurrence}
\alias{count_gene_recurrence}
\title{count of number of sites called per gene}
\usage{
count_gene_recurrence(de_novos)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}
}
\value{
data frame of de novos, but with an extra column indicating the
}
\description{
count of number of sites called per gene
}

