% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove_segdups.R
\name{exclude_segdups}
\alias{exclude_segdups}
\title{excludes de novo calls within segdup regions}
\usage{
exclude_segdups(de_novos, segdup_path)
}
\arguments{
\item{de_novos}{dataframe of candidate de novo calls}

\item{segdup_path}{path to file listing segmantal duplication regions
throughout the genome.}
}
\value{
dataframe of candidate de novo calls with sites in segmental
        duplications removed.
}
\description{
excludes de novo calls within segdup regions
}

