% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allele_counts.R
\name{extract_alt_and_ref_counts}
\alias{extract_alt_and_ref_counts}
\title{extract ALT and REF counts of forward and reverse reads for the members of
each trio}
\usage{
extract_alt_and_ref_counts(de_novos)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}
}
\value{
data frame of de novos, but with an extra columns for the read
        depths of forward and reverse reads for each member of the trio.
}
\description{
extract ALT and REF counts of forward and reverse reads for the members of
each trio
}

