% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_DNG.R
\name{fix_maf}
\alias{fix_maf}
\title{fix a few MAF values that are blank, or have lists of MAF values}
\usage{
fix_maf(de_novos)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}
}
\value{
vector of max allele frequency values for each site
}
\description{
fix a few MAF values that are blank, or have lists of MAF values
}

