% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/missing_symbols.R
\name{fix_missing_gene_symbols}
\alias{fix_missing_gene_symbols}
\title{adds gene symbols to variants lacking them, using a mupit function}
\usage{
fix_missing_gene_symbols(de_novos)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}
}
\value{
data frame of de novos, but with additional annotations for many
        variants previously lacking a HGNC symbol.
}
\description{
adds gene symbols to variants lacking them, using a mupit function
}

