% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_DNG.R
\name{get_allele_counts}
\alias{get_allele_counts}
\title{counts REF and ALT alleles in reads for candidate de novo events}
\usage{
get_allele_counts(vars, gene = FALSE)
}
\arguments{
\item{vars}{dataframe of de novo variants, all at a single site, or within a
single gene.}

\item{gene}{whether the set of variants of all within a single gene (rather
than at a site)}
}
\value{
list of forward reference, forward alternate.
}
\description{
counts REF and ALT alleles in reads for candidate de novo events
}

