% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set_filter_status.R
\name{get_filter_status}
\alias{get_filter_status}
\title{set flags for filtering, fail samples with strand bias < threshold, or any 2 of
 (i) both parents have ALTs
 (ii) site-specific parental alts < threshold,
 (iii) gene-specific parental alts < threshold, if > 1 sites called in gene}
\usage{
get_filter_status(de_novos)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}
}
\value{
vector of true/false for whether each variant passes the filters
}
\description{
set flags for filtering, fail samples with strand bias < threshold, or any 2 of
 (i) both parents have ALTs
 (ii) site-specific parental alts < threshold,
 (iii) gene-specific parental alts < threshold, if > 1 sites called in gene
}

