% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove_redundant_de_novos.R
\name{get_independent_de_novos}
\alias{get_independent_de_novos}
\title{remove de novos that are shared between multiple probands of a family or that
are recurrent in a single gene in a single person}
\usage{
get_independent_de_novos(de_novos, trios_path)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}

\item{trios_path}{path to a PED file listing all the trios}
}
\value{
dataframe with duplicated sites removed
}
\description{
remove de novos that are shared between multiple probands of a family or that
are recurrent in a single gene in a single person
}

