% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove_redundant_de_novos.R
\name{get_most_severe}
\alias{get_most_severe}
\title{get the most severe consequence from a list of VEP consequences}
\usage{
get_most_severe(consequences)
}
\arguments{
\item{consequences}{vector of VEP consequence strings}
}
\value{
single string for the most severe consequence
}
\description{
get the most severe consequence from a list of VEP consequences
}

