% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove_redundant_de_novos.R
\name{open_ddd_families}
\alias{open_ddd_families}
\title{opens a file defining the DDD individuals, and their families}
\usage{
open_ddd_families(path)
}
\arguments{
\item{path}{path to a PED file listing all the trios}
}
\value{
dataframe of families information
}
\description{
This gives us family information, so that we can identify siblings within a
family. We need to know this as sometimes de novo events recur within
families, so we have to remove all except for one, to prevent over-counting
the de novos.
}

