% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_DNG.R
\name{preliminary_filtering}
\alias{preliminary_filtering}
\title{run some preliminary filtering of de novos}
\usage{
preliminary_filtering(de_novos)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}
}
\value{
data frame of de novos, but where we have excluded sites with high
        allele frequency and where the variant appears in one or more parents.
}
\description{
We want to filter out de novos with high MAF, where they are not present in
the child VCF, or are present in the parental VCFs
}

