% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove_redundant_de_novos.R
\name{remove_within_person_recurrences}
\alias{remove_within_person_recurrences}
\title{find the de novos that are recurrent within a single individual in a
single gene. We shall treat these as a single de novo event. Prioritise
including the most severe event within a gene, then take the first variant
left after that.}
\usage{
remove_within_person_recurrences(de_novos)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}
}
\value{
dataframe with duplicated sites per gene removed for each individual
}
\description{
find the de novos that are recurrent within a single individual in a
single gene. We shall treat these as a single de novo event. Prioritise
including the most severe event within a gene, then take the first variant
left after that.
}

