% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_DNG.R
\name{site_strand_bias}
\alias{site_strand_bias}
\title{checks for strand bias per de novo site using the ref and alt counts}
\usage{
site_strand_bias(site)
}
\arguments{
\item{site}{dataframe of de novo variants, all at a single site, or within a
single gene.}
}
\value{
p-value for whether the forward or reverse are biased in the
       proportion of ref and alt alleles.
}
\description{
checks for strand bias per de novo site using the ref and alt counts
}

