% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set_filter_status.R
\name{subset_de_novos}
\alias{subset_de_novos}
\title{subset down to specific columns}
\usage{
subset_de_novos(de_novos)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}
}
\value{
data frame with only the pertinent columns included
}
\description{
subset down to specific columns
}

