% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_DNG.R
\name{test_genes}
\alias{test_genes}
\title{checks if the variants in a gene have more parental ALTs than expected}
\usage{
test_genes(de_novos)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}
}
\value{
p-value for whether the forward or reverse are biased in the
       proportion of ref and alt alleles within each gene.
}
\description{
checks if the variants in a gene have more parental ALTs than expected
}

