% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/site_deviations.R
\name{test_sites}
\alias{test_sites}
\title{tests each site for deviation from expected behaviour}
\usage{
test_sites(de_novos)
}
\arguments{
\item{de_novos}{dataframe of de novo variants}
}
\value{
p-value for whether the forward or reverse are biased in the
       proportion of ref and alt alleles.
}
\description{
tests each site for deviation from expected behaviour
}

